testing_rpi/requirements.txt
pyserial~=3.5
matplotlib~=3.6
numpy~=1.24
opencv-python~=4.6
